# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lab Playwright Kit — SaaS Parsing API",
    "description": "API для парсинга данных из браузера с антидетектом и авторизацией",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/parse": {
      "post": {
        "tags": [
          "Parsing"
        ],
        "summary": "Parse Url",
        "description": "Парсинг одного URL по схеме ниши.\n\nАвтоматически определяет нишу по URL если не указана.\nПоддерживает все 10 ниш: ecommerce, news, realty, medtech, jobs, auto, habr, vcru, twitter, telegram.",
        "operationId": "parse_url_api_v1_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parse/batch": {
      "post": {
        "tags": [
          "Parsing"
        ],
        "summary": "Parse Batch",
        "description": "Батч-парсинг нескольких URL.\n\nПараллельное выполнение с ограничением concurrency.\nЗадержка между запросами для избежания rate limiting.",
        "operationId": "parse_batch_api_v1_parse_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchParseRequest": {
        "properties": {
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Urls",
            "description": "Список URL"
          },
          "niche": {
            "type": "string",
            "title": "Niche",
            "description": "Ниша (пусто = автоопределение)",
            "default": ""
          },
          "timeout": {
            "type": "number",
            "title": "Timeout",
            "description": "Таймаут на каждый URL",
            "default": 30.0
          },
          "proxy_url": {
            "type": "string",
            "title": "Proxy Url",
            "description": "Прокси",
            "default": ""
          },
          "max_concurrency": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Concurrency",
            "description": "Параллельные воркеры",
            "default": 3
          },
          "delay_between": {
            "type": "number",
            "minimum": 0.0,
            "title": "Delay Between",
            "description": "Задержка между запросами (сек)",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "BatchParseRequest",
        "description": "Запрос на батч-парсинг."
      },
      "BatchParseResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "successful": {
            "type": "integer",
            "title": "Successful"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ParseResponse"
            },
            "type": "array",
            "title": "Results"
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "total",
          "successful",
          "failed",
          "results",
          "elapsed_seconds",
          "request_id"
        ],
        "title": "BatchParseResponse",
        "description": "Ответ батч-парсинга."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParseRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "URL для парсинга"
          },
          "niche": {
            "type": "string",
            "title": "Niche",
            "description": "Ниша (пусто = автоопределение)",
            "default": ""
          },
          "timeout": {
            "type": "number",
            "title": "Timeout",
            "description": "Таймаут в секундах",
            "default": 30.0
          },
          "proxy_url": {
            "type": "string",
            "title": "Proxy Url",
            "description": "Прокси (пусто = direct)",
            "default": ""
          },
          "wait_for": {
            "type": "string",
            "title": "Wait For",
            "description": "Селектор для ожидания перед парсингом",
            "default": ""
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Доп. данные"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ParseRequest",
        "description": "Запрос на парсинг одного URL."
      },
      "ParseResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "niche": {
            "type": "string",
            "title": "Niche"
          },
          "data": {
            "type": "object",
            "title": "Data"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "parse_time_ms": {
            "type": "number",
            "title": "Parse Time Ms"
          },
          "page_title": {
            "type": "string",
            "title": "Page Title"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "parsed_at": {
            "type": "string",
            "title": "Parsed At"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "url",
          "niche",
          "data",
          "confidence",
          "parse_time_ms",
          "page_title",
          "domain",
          "errors",
          "parsed_at",
          "request_id"
        ],
        "title": "ParseResponse",
        "description": "Ответ парсинга."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}